Fix: Fixed sonar issues
