+++ integrations/langchain4j/pom.xml
@@ -37,5 +37,11 @@
             <artifactId>langchain4j-core</artifactId>
             <version>${langchain4j.version}</version>
         </dependency>
+        <dependency>
+            <groupId>io.workm8.agui4j</groupId>
+            <artifactId>server</artifactId>
+            <version>0.0.1-SNAPSHOT</version>
+            <scope>compile</scope>
+        </dependency>
     </dependencies>
 </project>
