+++ integrations/spring-ai/pom.xml
@@ -61,6 +61,12 @@
             <version>6.2.9</version>
             <scope>compile</scope>
         </dependency>
+        <dependency>
+            <groupId>io.workm8.agui4j</groupId>
+            <artifactId>server</artifactId>
+            <version>0.0.1-SNAPSHOT</version>
+            <scope>compile</scope>
+        </dependency>
     </dependencies>
 
 </project>
