+++ integrations/spring-ai/src/main/java/io/workm8/agui4j/spring/ai/MessageMapper.java
@@ -5,7 +5,6 @@ import org.springframework.ai.chat.messages.*;
 import java.util.Map;
 import java.util.Objects;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
