+++ packages/core/src/test/java/io/workm8/agui4j/core/tool/ToolTest.java
@@ -8,7 +8,6 @@ import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
-import static org.assertj.core.util.Arrays.asList;
 
 @DisplayName("Tool")
 class ToolTest {
